# pkg/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Job Prep Backend API
  version: 1.0.0
  description: |
    API подготовки к собеседованиям, авторизации, тестов и асинхронной генерации по документам.
    Backend отправляет multipart-запрос на ASSESSMENT_SERVICE_URL/requests. Внешний сервис после
    обработки вызывает /api/v1/integrations/assistant/callback с X-Assessment-Callback-Secret.
servers:
  - url: /
    description: Текущий домен приложения
...
paths:
...
  /api/v1/integrations/assistant/callback:
    post:
      tags: [Integrations]
      summary: Callback внешнего сервиса генерации
      description: |
        Серверная точка возврата результата. Авторизуется общим секретом callback, а не API-ключом
        пользовательских интеграций. Для результата type=test тест атомарно сохраняется в аккаунт
        пользователя, которому принадлежит request_id.
      parameters:
        - in: header
          name: X-Assessment-Callback-Secret
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssessmentCallback" }
            examples:
              message:
                value:
                  request_id: 88e01249-2713-4bca-b0cc-72222483e8af
                  result:
                    type: message
                    message: Резюме получено. Для теста добавьте описание желаемой вакансии.
              test:
                value:
                  request_id: 88e01249-2713-4bca-b0cc-72222483e8af
                  result:
                    type: test
                    title: Подготовка к интервью Go-разработчика
                    description: Вопросы по опыту и навыкам из резюме
                    questions:
                      - text: Чем отличается map от sync.Map?
                        options: [Обычный map не потокобезопасен, Оба всегда потокобезопасны]
                        answer: Обычный map не потокобезопасен
              failed:
                value:
                  request_id: 88e01249-2713-4bca-b0cc-72222483e8af
                  status: failed
                  result:
                    type: message
                    message: Не удалось извлечь текст из резюме.
      responses:
        "200":
          description: Результат сохранён
        "204":
          description: Запрос помечен как неудачный
        "400":
          description: Некорректный результат генерации
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "401":
          description: Неверный callback secret
        "409":
          description: Запрос уже обработан или не существует
...
components:
  securitySchemes:
    integrationApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AssessmentCallback:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string, format: uuid }
        status:
          type: string
          enum: [failed]
          description: Укажи failed, если не удалось обработать документы; текст положи в result.message
        result:
          type: object
          properties:
            type: { type: string, enum: [message, test] }
            message: { type: string, description: Обязателен для type=message или status=failed }
            title: { type: string, description: Обязателен для type=test }
            description: { type: string }
            questions:
              type: array
              minItems: 1
              maxItems: 500
              items: { $ref: "#/components/schemas/QuestionInput" }
    QuestionInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1 }
        options: { type: array, items: {} }
        answer: {}
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...